Replaced 'create_function' by a closure.
A problem still remains; closures cannot be serialized. After some discussion, it seems like the best option is to forbid serializing the 'Mollie_API' class. I take it that serializing it in the test suite is a way of determining if through malicious attempts, one can retrieve the application secret, but other than that I see no actual use in it. So, preventing people to do so gets the job done as well.

Added PHP 7.2 to Travis.